Add new set of link and button helpers
These are clean implementations of that significantly simplify the
codebase and allow us to easily support features of the design
system, like opening links in new tabs. The new methods are:

* govuk_link_to
* govuk_button_to
* govuk_button_link_to
* govuk_link_classes
* govuk_button_classes
* govuk_breadcrumb_link_to

They drop support for Rails' legacy-style link building where the
action, controller and id arguments were passed in. Rails discourages
their use in the API docs[0]:

> Because it relies on url_for, link_to supports both older-style
> controller/action/id arguments and newer RESTful routes. Current Rails
> style favors RESTful routes whenever possible, so base your application
> on resources and use

[0] https://api.rubyonrails.org/v7.0/classes/ActionView/Helpers/UrlHelper.html#method-i-link_to
